--- a/src/Microsoft.PowerShell.Commands.Utility/commands/utility/Csv.cs
+++ b/src/Microsoft.PowerShell.Commands.Utility/commands/utility/Csv.cs
@@ -2,7 +2,8 @@
 // Licensed under the MIT License.
 
 using System.Collections.ObjectModel;
-using System.Text.RegularExpressions;
+//using System.Text.RegularExpressions;
+using Microsoft.VisualBasic.FileIO.TextFieldParser;
 
 namespace Microsoft.PowerShell.Commands
 {
@@ -29,16 +30,31 @@ namespace Microsoft.PowerShell.Commands
         /// </param>
         internal Collection<string> ParseCsv(string csv)
         {        
-            csv = csv.Trim();    
+            csv = csv.Trim(); 
+            Collection<string> result = new Collection<string>();
             if (csv.Length == 0 || csv[0] == '#')
             {
-                return new Collection<string>();
+                return result;
             }
 
+
+            var reader = new System.IO.StringReader(sampleText);
+            using (var parser = new Microsoft.VisualBasic.FileIO.TextFieldParser(reader))
+            {
+                parser.Delimiters = new string[] { "," };
+                parser.HasFieldsEnclosedInQuotes = true; // <--- !!!
+                string[] fields;
+                while ((fields = parser.ReadFields()) != null)
+                {
+                    result.Add(fields);
+                }
+            }
+            
             // regex expression, inspiration from
             // https://stackoverflow.com/questions/31118964/delimit-a-string-by-character-unless-within-quotation-marks-c-sharp
-            string[] elements = Regex.Split(csv, Delimiter + "|(\"[^\"]*\")");          
-            Collection<string> result = new Collection<string>(elements);          
+            // string[] elements = Regex.Split(csv, Delimiter + "|(\"[^\"]*\")"); 
+            // todo: exclude " in split
+            // Collection<string> result = new Collection<string>(elements);          
             return result;           
         }
     }