--- a/src/Microsoft.PowerShell.Commands.Utility/commands/utility/Csv.cs
+++ b/src/Microsoft.PowerShell.Commands.Utility/commands/utility/Csv.cs
@@ -2,8 +2,7 @@
 // Licensed under the MIT License.
 
 using System.Collections.ObjectModel;
-//using System.Text.RegularExpressions;
-//using System.IO;
+using System.IO;
 
 namespace Microsoft.PowerShell.Commands
 {
@@ -38,33 +37,38 @@ namespace Microsoft.PowerShell.Commands
                 return result;
             }
 
-            var reader = new System.IO.StringReader(csv);
+            var reader = new StringReader(csv);
             var tempString = string.Empty;
 
             // old implementation but now using the reader class
             while (reader.Peek() != -1) {
                 char nextChar = (char)reader.Read();
                 if (nextChar == Delimiter) {
+
                     //next character was delimiter found, so add string to collection
                     result.Add(tempString);
                     tempString = string.Empty;
                 } 
                 else if(nextChar == Quote) {
+
                     //next character was quote, so perform reading untill next quote and add it to tempString
                     bool isinQuotes = true;
                     while (reader.Peek() != -1 && isinQuotes) {
                         nextChar = (char)reader.Read();
                         if (nextChar == Quote) {
+
                             //exit quote found
                             isinQuotes = false;
                         } 
                         else {
+
                             // add to emptyString
                             tempString += nextChar;
                         }
                     }
                 }
             }
+
             // add last word if toAdd is not empty
             if(tempString != string.Empty) {
                 result.Add(tempString);