--- a/src/Microsoft.PowerShell.Commands.Utility/commands/utility/Csv.cs
+++ b/src/Microsoft.PowerShell.Commands.Utility/commands/utility/Csv.cs
@@ -74,8 +74,8 @@ namespace Microsoft.PowerShell.Commands
             if (tempString != string.Empty) {
                 result.Add(tempString);
             }
-            reader.Close();
-            //return new Collection<string>();      
+
+            reader.Close();    
             return result;           
         }
     }