@vexx32 2019-03-08
@SytzeAndr the `#` case is for when the CSV contains a type hint like it used to in Windows PowerShell -- CSVs generated without the `-NoTypeInformation` would have a one-line header with something like this:
```
PS Z:\> [pscustomobject]@{a = 1} | ConvertTo-Csv
#TYPE System.Management.Automation.PSCustomObject
"a"
"1"
```