@iSazonov 2019-03-09
@SytzeAndr Thanks for your contribution!

We are working on enhancing the cmdlet and have plans to improve performance.
I am concerned that your change affects performance and adds extra allocations. Did you measure?