@SytzeAndr 2019-03-09
@iSazonov I haven't tested it yet, but performance is one of the reasons I think refactoring this code would make sense. The old implementation is synchronous, but it should be possible to make this an asynchronous process instead of iterating over all the characters which should be faster. 

I haven't set up the testing environment locally (so therefore WIP), but when I do and the tests are passing I'll do a measure in terms of performance against the old implementation aswell.