@powercode 2019-03-09
Are you aware that this method is only called from the `Import-Alias` cmdlet?

The implementation of `Import-Csv` is in the file CSVCommands.cs, and the bulk of the work is done in the ImportCsvHelper class. That said, almost all of the cost of this operation is in creating the resulting objects. IIRC, the stream parsing was below 5%. And while the first rule of performance is to measure, in this case I would venture to speak from experience and say that using Regex is not the way to improve the performance.