@SytzeAndr 2019-03-12
some updates on this PR:

I've implemented the method by using the StringReader and StringBuilder. It currently passes all CI tests, so I assume my implementation is fine.

I assume use of the [Stringbuilder](
https://docs.microsoft.com/en-us/dotnet/api/system.text.stringbuilder?view=netcore-2.2) has a positive impact on performance since the `string` class builds a new string at each operation while the stringbuilder maintains it. 

I've reduced the complexity of this method according to CodeFactor (and is regarded as being a fixed issue to the repo). Although I still think it can be improved by splitting some things and might do so.

@powercode I used this PR aswell to get a bit used to contributing to powershell, and I chose to tackle an issue which could be fixed with some refactoring. Therefore, I didn't start with analyzing the bottlenecks since I had to set up things yet and wanted to start small by just refactoring some code. I agree with you in the sense that it would make more sense to analyze from a bit higher perspective to find the true bottlenecks. For a next PR, I'll try to do so, but a main reason for this PR is to just have something improved in powershell's code base, have my test environment set up, see what contributing to powershell is like, those sort of things. 

So the next step for me is to test it for performance against the old implementation, after that I'll remove the WIP tag (assuming performance is not worse). 
Is there some simple way of checking the performance/runtime by checking the CI runtime results for different branches, or is it better to do some local performance-based-tests myself of the different implementations and report them here?