@SytzeAndr 2019-03-13
I did some test and the performance seems to be improved, especially for large strings. I took the largest csv file from [here](https://sample-videos.com/download-sample-csv.php) called SampleCSVFile_53000kb.csv which contained 500.000 rows and is 53kb large. Then, I created some other very small csv file `very_small.csv` with the content `this, is, a, "very", small, file`. I did an experiment 100.000 times for the small file, and 100 times for the large file. The log is below.

For the large file, the performance seems to be twice as good (70s to 35s). For the small file, the performance is better, but not significantly (11.1s to 10.8s).

I'll remove the WIP tag and have this open for review.

```
file:csv_files/SampleCSVFile_53000kb.csv
iterations:100
useold:True
time elapsed:74262ms

file:csv_files/SampleCSVFile_53000kb.csv
iterations:100
version:False
time elapsed:35953ms

file:csv_files/very_small.csv
iterations:100000
useold:True
time elapsed:11080ms

file:csv_files/very_small.csv
iterations:100000
version:False
time elapsed:10759ms
````