@iSazonov 2019-03-13
@SytzeAndr Can you share the test script?