@SytzeAndr 2019-03-13
I used the following script. `CSVHelper` corresponds to my implementation, for `Csv_old` I copied code of the old implementation.

```
using System;
using Microsoft.PowerShell.Commands;
using System.IO;
using System.Diagnostics;

namespace csv_helper_new
{
    class Program
    {
        static void Main(string[] args)
        {
            bool useold = true;
            //args are files to read
            int iterations = 100;
            Stopwatch stopwatch = new Stopwatch();
            foreach (String a in args) {
                stopwatch.Reset();
                Console.WriteLine("file:"+a+"\n"+"iterations:"+iterations+"\nuseold:"+useold);
                stopwatch.Start();
                for (int i=0;i<iterations;i++) {
                    if (useold) {
                        Csv_old cSVHelper = new Csv_old(',');
                        cSVHelper.ParseCsv(File.ReadAllText(a)).ToString();
                    } else {
                        CSVHelper cSVHelper = new CSVHelper(',');
                        cSVHelper.ParseCsv(File.ReadAllText(a)).ToString();
                    }
                }
                stopwatch.Stop();
                Console.WriteLine("time elapsed:"+stopwatch.ElapsedTicks/1000+"ms");
            }            
        }
    }
}
```