@iSazonov 2019-03-13
@SytzeAndr Please clarify what do you want make faster? The `ParseCsv()` method is only used in Import-Alias cmdlet - do you really want to improve the cmdlet?