@LMiljak 2019-03-13
I think the main goal of this PR is to have the original code refactored (as the complexity of the method was quite high), without sacrificing performance. Making it faster was not necessarily the goal.